requirements.txt
python-telegram-bot==11.1.0
python-dotenv==0.10.1
pycurl==7.43.0
requests==2.9.1
urllib3==1.13.1

